+++ changelog.md
@@ -6,6 +6,7 @@ All notable changes to `laravel-alma-client` will be documented in this file.
 
 ### Fixed
 - Instanciate Alma Client take good config key
+- Client implementation
 
 ### Changed
 - Require Alma client ~1 instead of fixed ^1.8